+++ lib/pages/send_page.dart
@@ -6,15 +6,17 @@ import 'package:localsend_app/model/session_status.dart';
 import 'package:localsend_app/provider/device_info_provider.dart';
 import 'package:localsend_app/provider/network/send_provider.dart';
 import 'package:localsend_app/theme.dart';
-import 'package:localsend_app/util/sleep.dart';
 import 'package:localsend_app/widget/animations/initial_fade_transition.dart';
+import 'package:localsend_app/widget/animations/initial_slide_transition.dart';
 import 'package:localsend_app/widget/dialogs/error_dialog.dart';
 import 'package:localsend_app/widget/list_tile/device_list_tile.dart';
 import 'package:localsend_app/widget/responsive_list_view.dart';
 import 'package:routerino/routerino.dart';
 
 class SendPage extends ConsumerStatefulWidget {
-  const SendPage({Key? key}) : super(key: key);
+  final String sessionId;
+
+  const SendPage({required this.sessionId});
 
   @override
   ConsumerState<SendPage> createState() => _SendPageState();
@@ -24,24 +26,10 @@ class _SendPageState extends ConsumerState<SendPage> {
   Device? _myDevice;
   Device? _targetDevice;
 
-  // show name of sender after a delay
-  bool _showName = false;
-
-  @override
-  void initState() {
-    super.initState();
-
-    sleepAsync(800).then((value) {
-      if (mounted) {
-        setState(() => _showName = true);
-      }
-    });
-  }
-
   void _cancel() {
     // the state will be lost so we store them temporarily (only for UI)
     final myDevice = ref.read(deviceInfoProvider);
-    final sendState = ref.read(sendProvider);
+    final sendState = ref.read(sendProvider)[widget.sessionId];
     if (sendState == null) {
       return;
     }
@@ -50,12 +38,12 @@ class _SendPageState extends ConsumerState<SendPage> {
       _myDevice = myDevice;
       _targetDevice = sendState.target;
     });
-    ref.read(sendProvider.notifier).cancelSession();
+    ref.read(sendProvider.notifier).cancelSession(widget.sessionId);
   }
 
   @override
   Widget build(BuildContext context) {
-    final sendState = ref.watch(sendProvider);
+    final sendState = ref.watch(sendProvider)[widget.sessionId];
     if (sendState == null && _myDevice == null && _targetDevice == null) {
       return Scaffold(
         body: Container(),
@@ -81,11 +69,11 @@ class _SendPageState extends ConsumerState<SendPage> {
                     Expanded(
                       child: Column(
                         children: [
-                          Hero(
-                            tag: 'this-device',
+                          InitialSlideTransition(
+                            origin: const Offset(0, -1),
+                            duration: const Duration(milliseconds: 500),
                             child: DeviceListTile(
                               device: myDevice,
-                              thisDevice: !_showName,
                             ),
                           ),
                           const SizedBox(height: 20),