+++ assets/CHANGELOG.md
@@ -1,11 +1,15 @@
 ## 1.8.0 (2023-)
 
 - feat: add troubleshoot page
+- feat: add 2 buttons to receive history: open folder + delete history
 - feat: cleanup scan UI by hiding multiple network interfaces inside the scan button
+- feat: edit text message in selected files
+- feat(ex. iOS): pressing destination directory in progress page will open the directory
 - feat(android): share apk and install apk
 - feat(android): Android TV support
 - feat(windows): left click on tray icon opens app
 - feat(linux): add Control+Q shortcut to exit app
+- fix: handshake error in unencrypted mode
 - fix: also scan multicast when pressing on a subnet sync button
 - fix(android): missing app icon on Android 7
 - fix(android,ios): show error message when saving to gallery failed